+++ Backend/Market.Application/Modules/Meal/Queries/GetMealsByName/GetMealsByNameDto.cs
@@ -12,7 +12,7 @@ namespace Market.Application.Modules.Meal.Queries.GetMealsByName
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public double BasePrice { get; set; }
+        public decimal BasePrice { get; set; }
         public bool IsAvailable { get; set; }
         public bool IsFeatured { get; set; }
         public bool StockManaged { get; set; }
